requirements.txt
beautifulsoup4~=4.13.3
boto3~=1.34.162
botocore~=1.34.162
djangorestframework~=3.15.2
django~=5.1.6
gevent~=24.11.1
gunicorn~=23.0.0
huey-django-orm~=24.10.24.0
huey~=2.5.2
# lancedb 0.17.0 doesn't work with gunicorn :-(
lancedb~=0.16.0
langchain-community~=0.3.18
langchain-core~=0.3.40
langchain-google-genai~=2.0.11
langchain-openai~=0.3.7
langchain-text-splitters~=0.3.6
langchain~=0.3.19
markdown-it-py[plugins]~=3.0.0
pandas~=2.2.3
pdf2image~=1.17.0
pdfminer.six==20240706
pillow~=11.0.0
pyarrow~=19.0.1
python-dotenv~=1.0.1
s3fs~=2024.10.0
str2bool~=1.1
# Specify CPU-only torch to avoid dragging in Nvidia libs on Linux
--extra-index-url https://download.pytorch.org/whl/cpu
torch~=2.6.0
unstructured[pdf]~=0.16.20
